# RestaurantReservation.Backend/RestaurantReservation.Api/bin/Debug/netcoreapp3.1/Swaggerv1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Restaurant Reservation API v1",
    "description": "Restaurant Reservation API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Booking/Admin/ListCount": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Retrive the total number of members' pending reservation in database.",
        "operationId": "AdminGetCount",
        "responses": {
          "200": {
            "description": "Number of members' pending reservation record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving number of members' pending reservation record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/Booking/Admin/{status}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Get all members' bookings of a specific status.",
        "operationId": "GetAllBookingsOfStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Booking status, e.g. Pending, Expired, Complete",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Booking status, e.g. Pending, Expired, Complete",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All members' booking records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving members' booking records from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Booking/Details/BookingNo/{bookingNo}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Retrieve a specific booking details based on the Booking Number given.",
        "operationId": "GetByBookingNo",
        "parameters": [
          {
            "name": "bookingNo",
            "in": "path",
            "description": "Booking Number",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Booking Number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BookingDetailVM object that contains booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailVM"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving booking detail info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "400": {
            "description": "Booking Details Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Booking/Member/{userID}/{status}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Get all bookings of a specific status for a specific user.",
        "operationId": "GetAllBookingsByUserID",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "nullable": true
            }
          },
          {
            "name": "status",
            "in": "path",
            "description": "Order Status, eg. Pending, Complete, Expired",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Order Status, eg. Pending, Complete, Expired",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of BookingHistoryVM object that contains specific member's booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingHistoryVM"
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving booking detail info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Booking": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Add a new member booking to the database. This checks for number of seat avaiable for each session also.",
        "operationId": "Add",
        "requestBody": {
          "description": "MakeBookingVM object",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/MakeBookingVM"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/MakeBookingVM"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/MakeBookingVM"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/MakeBookingVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New Booking Object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "500": {
            "description": "Error Adding a member booking information into database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Booking/{bookingID}": {
      "put": {
        "tags": [
          "Booking"
        ],
        "summary": "Update an existing member's booking in the database.",
        "operationId": "Update",
        "parameters": [
          {
            "name": "bookingID",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Booking ID",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Booking Object Information",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Booking"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Deleting category successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid Booking ID message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No record found based on Booking ID given",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Updating a Booking object information from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationUser": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "joinedDate": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "isSubscriber": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "bookingID": {
            "type": "integer",
            "format": "int64"
          },
          "userID": {
            "type": "string",
            "nullable": true
          },
          "applicationUser": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "bookingDate": {
            "type": "string",
            "format": "date-time"
          },
          "checkIn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checkOut": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "bookingNo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "bookingStatus": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phoneNo": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "editedBy": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "editedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "BookingDetail": {
        "required": [
          "bookingID",
          "scheduleID"
        ],
        "type": "object",
        "properties": {
          "bookingDetailID": {
            "type": "integer",
            "format": "int64"
          },
          "bookingID": {
            "type": "integer",
            "format": "int64"
          },
          "booking": {
            "$ref": "#/components/schemas/Booking"
          },
          "scheduleID": {
            "type": "integer",
            "format": "int64"
          },
          "restaurantSchedule": {
            "$ref": "#/components/schemas/RestaurantSchedule"
          },
          "pax": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BookingDetailVM": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/Booking"
          },
          "bookingDetail": {
            "$ref": "#/components/schemas/BookingDetail"
          }
        },
        "additionalProperties": false
      },
      "BookingHistoryVM": {
        "type": "object",
        "properties": {
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            },
            "nullable": true
          },
          "bookingDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookingDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MakeBookingVM": {
        "type": "object",
        "properties": {
          "booking": {
            "$ref": "#/components/schemas/Booking"
          },
          "scheduleID": {
            "type": "integer",
            "format": "int64"
          },
          "pax": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: Bearer 12345abcdef",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}